+++ src/entities/Enemy.h
@@ -8,8 +8,8 @@ class Enemy : public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(double x READ x WRITE setX() NOTIFY xChanged)
-    Q_PROPERTY(double y READ y WRITE setY() NOTIFY yChanged)
+    Q_PROPERTY(float x READ x WRITE setX() NOTIFY xChanged)
+    Q_PROPERTY(float y READ y WRITE setY() NOTIFY yChanged)
     Q_PROPERTY(int m_index READ index WRITE setIndex NOTIFY indexChanged)
 
 public:
@@ -30,12 +30,12 @@ public:
         }
     }
 
-    double x()
+    float x()
     {
         return m_x;
     }
 
-    double y()
+    float y()
     {
         return m_y;
     }
@@ -65,10 +65,10 @@ signals:
     void indexChanged();
 
 private:
-    double m_x;
-    double m_y;
+    float m_x;
+    float m_y;
     int m_index;
-    double ySpeed;
+    float ySpeed;
     QTimer eTime;
 };
